+++ analysis.go
@@ -15,26 +15,37 @@ import (
 // Analyze is a reporter that processes stream and packet data for reports.
 // This must be in the Report pipeline *before* reporters that require it.
 type Analyze struct {
-	streams streams
-	packets packets
 }
 
 // report implements reporter
-func (y *Analyze) report(ctx context.Context, in <-chan any, out chan<- any,
+func (Analyze) report(ctx context.Context, in <-chan any, out chan<- any,
 	rw rwer) (err error) {
-	y.streams = newStreams()
-	y.packets = newPackets()
+	y := newAnalysis()
 	for d := range in {
 		out <- d
 		y.add(d)
 	}
 	y.analyze()
-	out <- *y
+	out <- y
 	return
 }
 
+// analysis contains the results of the Analyze reporter.
+type analysis struct {
+	streams streams
+	packets packets
+}
+
+// newAnalysis returns a new analysis.
+func newAnalysis() analysis {
+	return analysis{
+		newStreams(),
+		newPackets(),
+	}
+}
+
 // add adds a data item from the result stream.
-func (y *Analyze) add(a any) {
+func (y *analysis) add(a any) {
 	switch v := a.(type) {
 	case node.StreamInfo:
 		s := y.streams.analysis(v.Flow)
@@ -68,7 +79,7 @@ func (y *Analyze) add(a any) {
 }
 
 // analyze uses the collected data to calculate relevant metrics and stats.
-func (y *Analyze) analyze() {
+func (y *analysis) analyze() {
 	ss := y.streams.StartTime()
 	ps := y.packets.StartTime()
 	st := ss