+++ CHANGES.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Rename Test.During to Test.DuringDefault
 - Rename Test.Report to Test.AfterDefault
+- Make Analyze reporter concurrent-safe
 
 ## 0.4.0 - 2023-12-20